backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.19.0
pillow==10.1.0
numpy==1.26.4
python-multipart==0.0.6
scipy==1.11.4
pydantic-settings==2.0.3
pytest==7.4.3
httpx==0.25.2